Remove tage_sc_l
An adapted/simpler version may be added later.
The new version should be based on the tage example
and add the loop predictor and statistical corrector,
and being much shorter.
For learning purposes,
it would be better to care more about the general idea
and less about micro-optimizations.
The code from the championships is already there to be read.